publisher/requirements.txt
Flask==2.0.3
Pillow==9.0.1
requests==2.27.1
pika==1.2.0
waitress==2.1.1